Refactor codes.
